Move File Upload widgets to Exercise Downloads
